# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
    "/products/{id}/publish": {
      "post": {
        "operationId": "PublicationController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "publication"
        ]
      }
    },
    "/passport/{uuid}": {
      "get": {
        "operationId": "PublicPassportController_view",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
    "/passport/{uuid}/pdf": {
      "get": {
        "operationId": "PublicPassportController_pdf",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
    "/passport/{uuid}/qr.png": {
      "get": {
        "operationId": "PublicPassportController_qr",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
    "/passport/{uuid}/assets/{assetId}": {
      "get": {
        "operationId": "PublicPassportController_asset",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
    "/passport/{uuid}/view": {
      "post": {
        "operationId": "PublicPassportController_recordView",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewEventDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
    "/q/{uuid}": {
      "get": {
        "operationId": "PublicPassportController_resolveQr",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-passport"
        ]
      }
    },
...
  },
  "info": {
    "title": "Notarify Digital Product Passport API",
    "description": "Back office and anonymous API for the Notarify Digital Product Passport assessment.\n\nAuthentication uses a short-lived bearer access token plus an opaque rotating refresh\ncookie scoped to `/auth`. Obtain a token from `POST /auth/login`, then use the Authorize\nbutton with that token.\n\nAuthorization is server-authoritative: the access token carries no role claim, and the\nrole and company are re-read from PostgreSQL on every protected request. The Admin-only\nareas are Product deletion, Users, Settings and the audit log.\n\nThe anonymous public Passport surface requires no authentication at all.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local reviewer API origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /auth/login."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "PatchProductDto": {
        "type": "object",
        "properties": {}
      },
      "PublishProductDto": {
        "type": "object",
        "properties": {}
      },
      "ViewEventDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}